Nov 2

+++ operations/src/main/java/com/testingblaze/actionsfactory/elementfunctions/JavaScript.java
@@ -297,8 +297,8 @@ public final class JavaScript {
      * @return double y-axis
      * @author nauman.shahid
      */
-    public double getPageOffSetXAxis() {
-        return (double) js.executeScript("return window.pageXOffset");
+    public Long getPageOffSetXAxis() {
+        return  (Long)js.executeScript("return window.pageXOffset");
     }
 
     /**
@@ -307,8 +307,8 @@ public final class JavaScript {
      * @return double y-axis
      * @author nauman.shahid
      */
-    public double getPageOffSetYAxis() {
-        return (double) js.executeScript("return window.pageYOffset");
+    public Long getPageOffSetYAxis() {
+        return (Long) js.executeScript("return window.pageYOffset");
     }
 
     /**
